requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
google-api-core==1.31.0
google-auth==1.33.1
google-cloud==0.34.0
google-cloud-bigquery==2.22.1
google-cloud-bigquery-storage==2.6.1
google-cloud-core==1.7.1
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
libcst==0.3.19
loguru==0.5.3
mypy-extensions==0.4.3
numpy==1.19.5
packaging==21.0
pandas==1.1.5
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
win32-setctime==1.0.3
wincertstore==0.2
JayDeBeApi==1.2.3
psycopg2-binary==2.9.1
hdbcli==2.10.15
pytest
pre-commit
sqlfluff
apache-airflow
